requirements.txt
# Core dependencies
langchain>=0.1.0
langchain-community>=0.0.10

# PDF processing
pymupdf>=1.23.0
pymupdf4llm>=0.0.5

# Embeddings and ML
sentence-transformers>=2.2.2
torch>=2.0.0

# Vector store (PINNED to be compatible with server version 1.12.6)
# Server is v1.12.6, using latest 1.12.x client available (1.12.2)
qdrant-client==1.12.2

# BM25 search
rank-bm25>=0.2.2

# Tokenization
tiktoken>=0.5.0

# Evaluation
ragas>=0.1.0

# Web UI
streamlit>=1.29.0

# Caching
redis>=5.0.0

# Resilience
tenacity>=8.2.0
circuitbreaker>=2.0.0

# Observability
opentelemetry-api>=1.20.0
opentelemetry-sdk>=1.20.0
psutil>=5.9.0

# Configuration
pyyaml>=6.0

# HTTP client
httpx>=0.25.0

# Database
sqlite-utils>=3.35

# Utilities
tqdm>=4.66.0
python-dotenv>=1.0.0
fastapi>=0.109.2
uvicorn>=0.27.1

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# Quantization / Optimization
bitsandbytes>=0.40.0
accelerate>=0.20.0
tantivy>=0.22.0
